requirements-dev.txt
-r requirements.txt
pytest-django
django-stubs
twine
wheel
psycopg2-binary
